removed linux_terminal.py and moved contents to specific terminals
