+++ apps/platforms/linux/kde-konsole/kde-konsole.py
@@ -38,3 +38,18 @@ class app_actions:
     def window_open():
         actions.key('ctrl-shift-n')
 
+# this overwrites the unfitting parts of linux/edit.py
+@ctx.action_class('edit')
+class EditActions:
+    def page_down():
+        actions.key('shift-pagedown')
+    def page_up():
+        actions.key('shift-pageup')
+    def paste():
+        actions.key('ctrl-shift-v')
+    def copy():
+        actions.key('ctrl-shift-c')
+    def find(text: str):
+        actions.key('ctrl-shift-f')
+        if str:
+            actions.insert(text)
